+++ BCNotificationCentre/BCNotificationCentre.h
@@ -6,14 +6,7 @@
 //  Copyright © 2016 Bits & Co. All rights reserved.
 //
 
-#import <UIKit/UIKit.h>
+@import Foundation;
 
-//! Project version number for BCNotificationCentre.
 FOUNDATION_EXPORT double BCNotificationCentreVersionNumber;
-
-//! Project version string for BCNotificationCentre.
 FOUNDATION_EXPORT const unsigned char BCNotificationCentreVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <BCNotificationCentre/PublicHeader.h>
-
-