+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] - 2016-05-15
+### Added
+- Typed valued notifications
+- Travis integration
+- Documentation
+- Support for OSX, tvOS and watchOS
+
 ## [0.0.2] - 2016-05-14
 ### Fixed
 - Public init for `BCNotificationCentre` and `BCNotification`